requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
Flask-Login
pywebview
pyinstaller
